docs: mention animated coffee icon
